Factor out repeated duckdb initialization logic
